+++ Barometer.h
@@ -8,7 +8,11 @@
  * Create Time:
  * Change Log :
  *
+ * loovee 9-24-2014
+ * Change all int to short, all unsigned int to unsigned short to fit some 32bit system
+ *
  * The MIT License (MIT)
+ *
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -36,33 +40,35 @@
 
 const unsigned char OSS = 0;
 #define BMP085_ADDRESS 0x77
+
 class Barometer
 {
-public:
-  void init(void);
-  long PressureCompensate;
-  float bmp085GetTemperature(unsigned int ut);
-  long bmp085GetPressure(unsigned long up);
-  float calcAltitude(float pressure);
-  unsigned int bmp085ReadUT(void);
-  unsigned long bmp085ReadUP(void);
+    public:
+    void init(void);
+    long PressureCompensate;
+    float bmp085GetTemperature(unsigned short ut);
+    long bmp085GetPressure(unsigned long up);
+    float calcAltitude(float pressure);
+    unsigned short bmp085ReadUT(void);
+    unsigned long bmp085ReadUP(void);
 
-private:
-  int ac1;
-  int ac2;
-  int ac3;
-  unsigned int ac4;
-  unsigned int ac5;
-  unsigned int ac6;
-  int b1;
-  int b2;
-  int mb;
-  int mc;
-  int md;
-  char bmp085Read(unsigned char address);
-  int bmp085ReadInt(unsigned char address);
-  void writeRegister(int deviceAddress, byte address, byte val);
-  int readRegister(int deviceAddress, byte address);
+    private:
+
+    short ac1;
+    short ac2;
+    short ac3;
+    unsigned short ac4;
+    unsigned short ac5;
+    unsigned short ac6;
+    short b1;
+    short b2;
+    short mb;
+    short mc;
+    short md;
+    char bmp085Read(unsigned char address);
+    short bmp085ReadInt(unsigned char address);
+    void writeRegister(short deviceAddress, byte address, byte val);
+    short readRegister(short deviceAddress, byte address);
 };
 
 #endif
